requirements.txt
streamlit
fastapi
uvicorn[standard]
langchain
langchain-anthropic
langchain-core
langgraph
pydantic
faiss-cpu
sentence-transformers
pandas
scikit-learn
numpy
numba
RestrictedPython
sqlalchemy
python-dotenv
requests
joblib
structlog
aiofiles
ydata-profiling
pytest
httpx
python-multipart
matplotlib
typing-extensions